Fixing the template variable names, splitting out utils.ts into separate files, removing react-is usage which wasn't in the package.json, no need to pull an external dependency to check if we're a forwardref.
