+++ packages/react-compose/etc/react-compose.api.md
@@ -10,19 +10,19 @@ import * as React from 'react';
 export type ClassDictionary = Record<string, string>;
 
 // @public (undocumented)
-export interface ComponentWithAs<E extends React.ElementType = 'div', P = {}> extends React.FunctionComponent {
+export interface ComponentWithAs<TElementType extends React.ElementType = 'div', TProps = {}> extends React.FunctionComponent {
     // (undocumented)
-    <EE extends React.ElementType = E>(props: Omit<PropsOfElement<EE>, 'as' | keyof P> & {
-        as?: EE;
-    } & P): JSX.Element | null;
+    <TExtendedElementType extends React.ElementType = TElementType>(props: Omit<PropsOfElement<TExtendedElementType>, 'as' | keyof TProps> & {
+        as?: TExtendedElementType;
+    } & TProps): JSX.Element | null;
     // (undocumented)
-    defaultProps?: Partial<P & {
-        as: E;
+    defaultProps?: Partial<TProps & {
+        as: TElementType;
     }>;
     // (undocumented)
     displayName?: string;
     // (undocumented)
-    propTypes?: React.WeakValidationMap<P> & {
+    propTypes?: React.WeakValidationMap<TProps> & {
         as: React.Requireable<string | ((props: any, context?: any) => any) | (new (props: any, context?: any) => any)>;
     };
 }
@@ -31,43 +31,43 @@ export interface ComponentWithAs<E extends React.ElementType = 'div', P = {}> ex
 export function compose<ElementType extends React.ElementType, InputProps, InputStylesProps, ParentProps, ParentStylesProps>(input: Input<ElementType, InputProps>, inputOptions?: ComposeOptions<InputProps, InputStylesProps, ParentProps, ParentStylesProps>): ComponentWithAs<ElementType, InputProps & ParentProps>;
 
 // @public (undocumented)
-export type ComposedComponent<P = {}> = React.FunctionComponent<P> & {
+export type ComposedComponent<TProps = {}> = React.FunctionComponent<TProps> & {
     fluentComposeConfig: ComposePreparedOptions;
 };
 
 // @public (undocumented)
-export type ComposeOptions<InputProps = {}, InputStylesProps = {}, ParentProps = {}, ParentStylesProps = {}> = {
+export type ComposeOptions<TInputProps = {}, TInputStylesProps = {}, TParentProps = {}, TParentStylesProps = {}> = {
     className?: string;
     classes?: ClassDictionary;
     displayName?: string;
-    mapPropsToStylesProps?: (props: ParentStylesProps & InputProps) => InputStylesProps;
-    handledProps?: (keyof InputProps | 'as')[];
+    mapPropsToSlotProps?: (props: TParentProps & TInputProps) => Record<string, object>;
+    handledProps?: (keyof TInputProps | 'as')[];
     overrideStyles?: boolean;
     slots?: Record<string, React.ElementType>;
-    mapPropsToSlotProps?: (props: ParentProps & InputProps) => Record<string, object>;
-    shorthandConfig?: ShorthandConfig<ParentProps & InputProps>;
+    mapPropsToStylesProps?: (props: TParentStylesProps & TInputProps) => TInputStylesProps;
+    shorthandConfig?: ShorthandConfig<TParentProps & TInputProps>;
 };
 
 // @public (undocumented)
-export type ComposePreparedOptions<Props = {}> = {
+export type ComposePreparedOptions<TElementType extends React.ElementType = 'div', TProps = {}, TState = TProps> = {
     className: string;
     classes: ClassDictionary;
     displayName: string;
     displayNames: string[];
     mapPropsToStylesPropsChain: ((props: object) => object)[];
     render: ComposeRenderFunction;
-    handledProps: (keyof Props)[];
+    handledProps: (keyof TProps)[];
     overrideStyles: boolean;
     slots: Record<string, React.ElementType> & {
         __self: React.ElementType;
     };
-    mapPropsToSlotPropsChain: ((props: Props) => Record<string, object>)[];
-    resolveSlotProps: <P>(props: P) => Record<string, object>;
-    shorthandConfig: ShorthandConfig<Props>;
+    mapPropsToSlotPropsChain: ((props: TProps) => Record<string, object>)[];
+    resolveSlotProps: <P>(props: TProps) => Record<string, object>;
+    shorthandConfig: ShorthandConfig<TProps>;
 };
 
 // @public (undocumented)
-export type ComposeRenderFunction<T extends React.ElementType = 'div', P = {}> = (props: P, ref: React.Ref<T extends keyof HTMLElementTagNameMap ? HTMLElementTagNameMap[T] : T>, composeOptions: ComposePreparedOptions) => React.ReactElement | null;
+export type ComposeRenderFunction<TElementType extends React.ElementType = 'div', TProps = {}> = (props: TProps, ref: React.Ref<TElementType extends keyof HTMLElementTagNameMap ? HTMLElementTagNameMap[TElementType] : TElementType>, composeOptions: ComposePreparedOptions) => React.ReactElement | null;
 
 // @public
 export const createClassResolver: (classes: Record<string, string>) => (state: Record<string, any>, slots: Record<string, any>) => Record<string, string>;
@@ -76,12 +76,12 @@ export const createClassResolver: (classes: Record<string, string>) => (state: R
 export type GenericDictionary = Record<string, any>;
 
 // @public (undocumented)
-export type Input<T extends React.ElementType = 'div', P = {}> = InputComposeComponent<P> | ComposeRenderFunction<T, P & {
+export type Input<TElementType extends React.ElementType = 'div', TProps = {}> = InputComposeComponent<TProps> | ComposeRenderFunction<TElementType, TProps & {
     as?: React.ElementType;
 }>;
 
 // @public (undocumented)
-export type InputComposeComponent<P = {}> = React.FunctionComponent<P> & {
+export type InputComposeComponent<TProps = {}> = React.FunctionComponent<TProps> & {
     fluentComposeConfig?: ComposePreparedOptions;
 };
 
@@ -89,13 +89,13 @@ export type InputComposeComponent<P = {}> = React.FunctionComponent<P> & {
 export type PropsOfElement<E extends keyof JSX.IntrinsicElements | React.JSXElementConstructor<any>> = JSX.LibraryManagedAttributes<E, React.ComponentPropsWithRef<E>>;
 
 // @public (undocumented)
-export interface ShorthandConfig<P> {
+export interface ShorthandConfig<TProps> {
     // (undocumented)
     allowsJSX?: boolean;
     // (undocumented)
-    mappedArrayProp?: keyof P;
+    mappedArrayProp?: keyof TProps;
     // (undocumented)
-    mappedProp?: keyof P;
+    mappedProp?: keyof TProps;
 }
